move debug things into testing
